Merge master into release

+++ src/undo/undo_manager.cpp
@@ -41,7 +41,7 @@
 
 namespace OpenOrienteering {
 
-Q_STATIC_ASSERT(UndoManager::max_undo_steps < std::numeric_limits<int>::max());
+Q_STATIC_ASSERT(UndoManager::max_undo_steps < static_cast<unsigned int>(std::numeric_limits<int>::max()));
 
 
 // ### UndoManager::State ###
